Normalize int values when port is defined within the spec
When port values as provided as a string, cephadm fails to
apply the generated spec: for instance, if we have something
like:

---
placement:
  hosts:
  - standalone.localdomain
service_id: grafana
service_name: grafana
service_type: grafana
spec:
  port: '3100'

it will fail, because the port value is expected as an integer.
For this reason, we're adding the normalize_spec function to
the ceph_mkspec module in order to be able to prevent this kind
of failures when ansible evaluates everything as a string.
By doing this, the resulting dictionary will be something like:

---
placement:
  hosts:
  - standalone.localdomain
service_id: grafana
service_name: grafana
service_type: grafana
spec:
  port: 3100

which is acceptable.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>